Update @faker-js/faker dependency to version 9.0.0, refactor FakerSchemaEditor to handle loading state, and improve type definitions for FakerSchemaMapping.
