+++ packages/compass-collection/src/components/mock-data-generator-modal/mock-data-generator-modal.tsx
@@ -11,7 +11,6 @@ import {
   Modal,
   ModalFooter,
   spacing,
-  SpinLoaderWithLabel,
 } from '@mongodb-js/compass-components';
 
 import { type MockDataGeneratorState, MockDataGeneratorStep } from './types';
@@ -43,12 +42,6 @@ const namespaceStyles = css({
   marginBottom: spacing[400],
 });
 
-const schemaEditorLoaderStyles = css({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
-
 interface Props {
   isOpen: boolean;
   onClose: () => void;
@@ -78,28 +71,13 @@ const MockDataGeneratorModal = ({
       case MockDataGeneratorStep.SCHEMA_CONFIRMATION:
         return <RawSchemaConfirmationScreen />;
       case MockDataGeneratorStep.SCHEMA_EDITOR:
-        {
-          if (fakerSchemaGenerationState.status === 'in-progress') {
-            return (
-              <div
-                data-testid="faker-schema-editor-loader"
-                className={schemaEditorLoaderStyles}
-              >
-                <SpinLoaderWithLabel progressText="Processing Documents..." />
-              </div>
-            );
-          }
-          if (fakerSchemaGenerationState.status === 'completed') {
-            return (
-              <FakerSchemaEditorScreen
-                isSchemaConfirmed={isSchemaConfirmed}
-                onSchemaConfirmed={setIsSchemaConfirmed}
-                fakerMappings={fakerSchemaGenerationState.fakerSchema}
-              />
-            );
-          }
-        }
-        break;
+        return (
+          <FakerSchemaEditorScreen
+            isSchemaConfirmed={isSchemaConfirmed}
+            onSchemaConfirmed={setIsSchemaConfirmed}
+            fakerSchemaGenerationState={fakerSchemaGenerationState}
+          />
+        );
       case MockDataGeneratorStep.DOCUMENT_COUNT:
         return <></>; // TODO: CLOUDP-333856
       case MockDataGeneratorStep.PREVIEW_DATA:
@@ -110,8 +88,7 @@ const MockDataGeneratorModal = ({
   }, [currentStep, fakerSchemaGenerationState, isSchemaConfirmed]);
 
   const isNextButtonDisabled =
-    currentStep === MockDataGeneratorStep.SCHEMA_EDITOR &&
-    (!isSchemaConfirmed || fakerSchemaGenerationState.status === 'in-progress');
+    currentStep === MockDataGeneratorStep.SCHEMA_EDITOR && !isSchemaConfirmed;
 
   const handleNextClick = () => {
     if (currentStep === MockDataGeneratorStep.GENERATE_DATA) {
